producer/requirements.txt
pulsar-client
flask